integrate a common header component on all pages (#32)
