+++ package.json
@@ -107,7 +107,7 @@
     "neutrino": "^9.3.0",
     "node-sass": "^4.13.1",
     "npm": "^6.14.8",
-    "prettier": "^2.0.5",
+    "prettier": "^2.1.2",
     "ts-loader": "^6.2.2",
     "tslint": "^6.1.2",
     "typescript": "^3.8.3"
