+++ webapp/src/containers/ConsolePage/index.tsx
@@ -3,6 +3,7 @@ import { I18n } from 'react-redux-i18n';
 import { Helmet } from 'react-helmet';
 import { Row, Col } from 'reactstrap';
 import Console from './ConsoleComponent';
+import Header from '../HeaderComponent';
 
 const ConsolePage: React.FunctionComponent = () => {
   return (
@@ -12,9 +13,9 @@ const ConsolePage: React.FunctionComponent = () => {
           {I18n.t('containers.console.consolePage.consoleDefiClient')}
         </title>
       </Helmet>
-      <header className='header-bar'>
+      <Header>
         <h1>{I18n.t('containers.console.consolePage.console')}</h1>
-      </header>
+      </Header>
       <div className='content'>
         <section>
           <Row>
