+++ webapp/src/containers/HelpPage/index.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { Helmet } from 'react-helmet';
 import { I18n } from 'react-redux-i18n';
+import Header from '../HeaderComponent';
 
 const HelpPage = () => {
   return (
@@ -8,9 +9,9 @@ const HelpPage = () => {
       <Helmet>
         <title>{I18n.t('containers.helpPage.title')}</title>
       </Helmet>
-      <header className='header-bar'>
+      <Header>
         <h1>{I18n.t('containers.helpPage.help')}</h1>
-      </header>
+      </Header>
       <div className='content'>
         <section>{I18n.t('containers.helpPage.section')}</section>
       </div>
