+++ webapp/src/containers/TokensPage/components/MintToken/index.tsx
@@ -29,6 +29,7 @@ import Spinner from '../../../../components/Svg/Spinner';
 import { connect } from 'react-redux';
 import { mintToken } from '../../reducer';
 import { isEmpty } from 'lodash';
+import Header from '../../../HeaderComponent';
 
 interface RouteParams {
   id?: string;
@@ -113,7 +114,7 @@ const MintToken: React.FunctionComponent<MintTokenProps> = (
       <Helmet>
         <title>{I18n.t('containers.tokens.mintToken.title')}</title>
       </Helmet>
-      <header className='header-bar'>
+      <Header>
         <Button
           to={`${TOKENS_PATH}/${id}/${hash}`}
           tag={NavLink}
@@ -130,7 +131,7 @@ const MintToken: React.FunctionComponent<MintTokenProps> = (
           &nbsp;
           {id}
         </h1>
-      </header>
+      </Header>
       <div className='content'>
         <section>
           <Form>
