+++ webapp/src/containers/TokensPage/components/TokenInfo/index.tsx
@@ -36,6 +36,7 @@ import {
 } from '../../../../constants';
 import { ITokenResponse } from '../../../../utils/interfaces';
 import { getIcon } from '../../../../utils/utility';
+import Header from '../../../HeaderComponent';
 
 interface RouteParams {
   id?: string;
@@ -135,7 +136,7 @@ const TokenInfo: React.FunctionComponent<TokenInfoProps> = (
       <Helmet>
         <title>{I18n.t('containers.tokens.tokensPage.title')}</title>
       </Helmet>
-      <header className='header-bar'>
+      <Header>
         <Button
           to={TOKENS_PATH}
           tag={RRNavLink}
@@ -178,7 +179,7 @@ const TokenInfo: React.FunctionComponent<TokenInfoProps> = (
             </DropdownMenu>
           </UncontrolledDropdown>
         )}
-      </header>
+      </Header>
       {tokenInfo.minted === 0 && tokenInfo.hash !== '0' && tokenInfo.ismine && (
         <div className={`${styles.mintAlert} m-5`}>
           <span>{I18n.t('containers.tokens.tokenInfo.notMintedAlert')}</span>
