+++ webapp/src/containers/WalletPage/components/CreateWallet/CreateNewWallet/index.tsx
@@ -17,6 +17,7 @@ import {
 
 import { WALLET_PAGE_PATH } from '../../../../../constants';
 import styles from '../CreateWallet.module.scss';
+import Header from '../../../../HeaderComponent';
 
 interface CreateNewWalletProps {
   mnemonicObj: any;
@@ -42,7 +43,7 @@ const CreateNewWallet: React.FunctionComponent<CreateNewWalletProps> = (
       <Helmet>
         <title>{I18n.t('containers.wallet.createNewWalletPage.title')}</title>
       </Helmet>
-      <header className='header-bar'>
+      <Header>
         <Button
           to={WALLET_PAGE_PATH}
           tag={NavLink}
@@ -57,7 +58,7 @@ const CreateNewWallet: React.FunctionComponent<CreateNewWalletProps> = (
         <h1 className={classnames({ 'd-none': false })}>
           {I18n.t('containers.wallet.createNewWalletPage.createANewWallet')}
         </h1>
-      </header>
+      </Header>
       <div className='content'>
         <section>
           <p>
