+++ webapp/src/containers/WalletPage/index.tsx
@@ -22,7 +22,8 @@ import { startUpdateApp, openBackupWallet } from '../PopOver/reducer';
 import { WALLET_SEND_PATH, WALLET_RECEIVE_PATH } from '../../constants';
 import { getIcon, getAmountInSelectedUnit } from '../../utils/utility';
 import styles from './WalletPage.module.scss';
-import Badge from '../../components/Badge';
+// import Badge from '../../components/Badge';
+import Header from '../HeaderComponent';
 
 interface WalletPageProps extends RouteComponentProps {
   unit: string;
@@ -30,9 +31,6 @@ interface WalletPageProps extends RouteComponentProps {
   pendingBalance: string;
   fetchInstantBalanceRequest: () => void;
   fetchInstantPendingBalanceRequest: () => void;
-  updateAvailableBadge: boolean;
-  startUpdateApp: () => void;
-  openBackupWallet: () => void;
   blockChainInfo: any;
 }
 
@@ -49,10 +47,8 @@ const WalletPage: React.FunctionComponent<WalletPageProps> = (
     fetchInstantBalanceRequest,
     unit,
     fetchInstantPendingBalanceRequest,
-    updateAvailableBadge,
-    startUpdateApp,
-    openBackupWallet,
-    history,
+    walletBalance,
+    pendingBalance,
   } = props;
 
   useEffect(() => {
@@ -65,14 +61,8 @@ const WalletPage: React.FunctionComponent<WalletPageProps> = (
     };
   }, []);
 
-  const openUpdatePopUp = () => {
-    openBackupWallet();
-    startUpdateApp();
-  };
-
   let balanceRefreshTimerID;
   let pendingBalRefreshTimerID;
-  const { walletBalance, pendingBalance } = props;
   const [refreshBalance, setRefreshBalance] = useState(false);
   const [pendingRefreshBalance, setPendingRefreshBalance] = useState(false);
 
@@ -81,21 +71,21 @@ const WalletPage: React.FunctionComponent<WalletPageProps> = (
       <Helmet>
         <title>{I18n.t('containers.wallet.walletPage.wallet')}</title>
       </Helmet>
-      <header className='header-bar'>
-          <Button
-            to={`${WALLET_TOKENS_PATH}?value=${getAmountInSelectedUnit(
-              walletBalance,
-              unit
-            )}&unit=${unit}`}
-            tag={RRNavLink}
-            color='link'
-            className='header-bar-back'
-          >
-            <MdArrowBack />
-            <span className='d-lg-inline'>
-              {I18n.t('containers.wallet.walletPage.tokens')}
-            </span>
-          </Button>
+      <Header>
+        <Button
+          to={`${WALLET_TOKENS_PATH}?value=${getAmountInSelectedUnit(
+            walletBalance,
+            unit
+          )}&unit=${unit}`}
+          tag={RRNavLink}
+          color='link'
+          className='header-bar-back'
+        >
+          <MdArrowBack />
+          <span className='d-lg-inline'>
+            {I18n.t('containers.wallet.walletPage.tokens')}
+          </span>
+        </Button>
         <div className='d-flex'>
           <img src={getIcon(tokenSymbol)} height={'30px'} width={'30px'} />
           &nbsp;
@@ -105,14 +95,6 @@ const WalletPage: React.FunctionComponent<WalletPageProps> = (
             {I18n.t('containers.wallet.walletPage.wallet')}
           </h1>
         </div>
-        {updateAvailableBadge && (
-          <Badge
-            baseClass='update-available'
-            outline
-            onClick={openUpdatePopUp}
-            label={I18n.t('containers.wallet.walletPage.updateAvailableLabel')}
-          />
-        )}
         <ButtonGroup>
           <Button
             to={
@@ -145,7 +127,7 @@ const WalletPage: React.FunctionComponent<WalletPageProps> = (
             </span>
           </Button>
         </ButtonGroup>
-      </header>
+      </Header>
       <div className='content'>
         <section>
           <Row>
@@ -209,13 +191,11 @@ const mapStateToProps = (state) => {
     settings: {
       appConfig: { unit },
     },
-    popover: { updateAvailableBadge },
   } = state;
   return {
     unit,
     walletBalance,
     pendingBalance,
-    updateAvailableBadge,
     blockChainInfo,
   };
 };
